+++ src/components/Setup/Setup.js
@@ -167,10 +167,10 @@ function Setup(props) {
                       {/* <span onClick={handleTravel} style={{fontWeight:'bold', cursor:'pointer'}}>Click here...</span> (Alpha Release) */}
         </Typography> }
         <span onClick={() => setPage('Travel')}>
-        {/* <Typography color="textSecondary" variant="h2" component="h2" align="left" 
+        <Typography color="textSecondary" variant="h2" component="h2" align="left" 
       style={{fontWeight:'bold', fontSize:'1rem', padding:'10px 24px', color: 'rgb(3, 155, 229)', marginBottom:0}} gutterBottom>
-                  Travellers Click here ...
-      </Typography> */}
+                  Travellers click here ... ( in α state)
+      </Typography>
       </span>
         <Stepper activeStep={activeStep} orientation="vertical" >
           {steps.map((label, index) => (