+++ src/components/Timer/LocationText.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import {useCurrentLocation} from '../../hooks/api-hooks';
+// import {useCurrentLocation} from '../../hooks/api-hooks';
 
-const Locationtext = () => {
-    const [currentloc] = useCurrentLocation({lat:"52.31406610552598",  lon:"4.946411339519716"});
-    console.log('%c LOCATIOn'+currentloc, 'color:purple')
+const Locationtext = ({location}) => {
+    // const [currentloc] = useCurrentLocation({lat:"52.31406610552598",  lon:"4.946411339519716"});
+    // console.log('%c LOCATIOn'+currentloc, 'color:purple')
     return (
         <>
-            <span className="locationText">@ {currentloc.data}</span>
+            <span className="locationText">@ {location}</span>
         </>
     )
 }